Change to float.

+++ sql/updates/world/3.3.5/9999_99_99_99_world.sql
@@ -1,3 +1,3 @@
 --
-ALTER TABLE `creature_template_addon` ADD COLUMN `visibilityRange` INT(3) UNSIGNED DEFAULT 0 NOT NULL AFTER `emote`;
-ALTER TABLE `gameobject_template_addon` ADD COLUMN `visibilityRange` INT(3) UNSIGNED DEFAULT 0 NOT NULL AFTER `maxgold`;
+ALTER TABLE `creature_template_addon` ADD COLUMN `visibilityRange` FLOAT(7, 4) UNSIGNED DEFAULT 0 NOT NULL AFTER `emote`;
+ALTER TABLE `gameobject_template_addon` ADD COLUMN `visibilityRange` FLOAT(7, 4) UNSIGNED DEFAULT 0 NOT NULL AFTER `maxgold`;
